requirements.txt
celery==5.0.5
django==2.2.16
django-auditlog==0.4.7
django-filter==2.4.0
django-tenants==3.2.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pytz==2020.5
psycopg2-binary==2.8.6
pycodestyle==2.6.0